refactor: Ladder 클래스 분리
- LadderGame 내부에서 Ladder를 만드는 로직을 Ladder 클래스에서 처리하도록 따로 분리하였습니다.
